+++ python/ray/worker.py
@@ -363,10 +363,14 @@ class Worker(object):
                         logger.warning(warning_message)
 
     def put_object(self, object_id, value):
-        """Put value in the local object store with object id objectid.
+        """Put value in the local object store with object id `objectid`.
 
-        This assumes that the value for objectid has not yet been placed in the
-        local object store.
+        This assumes that the value for `objectid` has not yet been placed in
+        the local object store. If the plasma store is full, the worker will
+        automatically retry up to DEFAULT_PUT_OBJECT_RETRIES times. Each
+        retry will delay for an exponentially doubling amount of time,
+        starting with DEFAULT_PUT_OBJECT_DELAY. After this, exception
+        will be raised.
 
         Args:
             object_id (object_id.ObjectID): The object ID of the value to be
@@ -374,10 +378,9 @@ class Worker(object):
             value: The value to put in the object store.
 
         Raises:
-            Exception: An exception is raised if the attempt to store the
-                object fails. This can happen if there is already an object
-                with the same ID in the object store or if the object store is
-                full.
+            plasma.PlasmaStoreFull: This is raised if the attempt to store the
+                object fails because the object store is full even after
+                multiple retries.
         """
         # Make sure that the value is not an object ID.
         if isinstance(value, ObjectID):
@@ -388,35 +391,11 @@ class Worker(object):
                 "do this, you can wrap the ray.ObjectID in a list and "
                 "call 'put' on it (or return it).")
 
-        def try_store_and_register():
-            # Serialize and put the object in the object store.
-            try:
-                self.store_and_register(object_id, value)
-            except pyarrow.plasma.PlasmaObjectExists:
-                # The object already exists in the object store, so there is no
-                # need to add it again. TODO(rkn): We need to compare hashes
-                # and make sure that the objects are in fact the same. We also
-                # should return an error code to caller instead of printing a
-                # message.
-                logger.info("The object with ID {} already exists "
-                            "in the object store.".format(object_id))
-            except TypeError:
-                # TypeError can happen because one of the members of the object
-                # may not be serializable for cloudpickle. So we need
-                # these extra fallbacks here to start from the beginning.
-                # Hopefully the object could have a `__reduce__` method.
-                register_custom_serializer(type(value), use_pickle=True)
-                warning_message = ("WARNING: Serializing the class {} failed, "
-                                   "falling back to cloudpickle.".format(
-                                       type(value)))
-                logger.warning(warning_message)
-                self.store_and_register(object_id, value)
-
         delay = ray_constants.DEFAULT_PUT_OBJECT_DELAY
         for attempt in reversed(
                 range(ray_constants.DEFAULT_PUT_OBJECT_RETRIES)):
             try:
-                try_store_and_register()
+                self.try_store_and_register(object_id, value)
                 break
             except pyarrow.plasma.PlasmaStoreFull as plasma_exc:
                 if attempt:
@@ -426,6 +405,36 @@ class Worker(object):
                     delay *= 2
                 else:
                     raise plasma_exc
+
+    def _try_store_and_register(self, object_id, value):
+        """Wraps `store_and_register` with cases for existence and pickling.
+
+        Args:
+            object_id (object_id.ObjectID): The object ID of the value to be
+                put.
+            value: The value to put in the object store.
+        """
+        try:
+            self.store_and_register(object_id, value)
+        except pyarrow.plasma.PlasmaObjectExists:
+            # The object already exists in the object store, so there is no
+            # need to add it again. TODO(rkn): We need to compare hashes
+            # and make sure that the objects are in fact the same. We also
+            # should return an error code to caller instead of printing a
+            # message.
+            logger.info("The object with ID {} already exists "
+                        "in the object store.".format(object_id))
+        except TypeError:
+            # TypeError can happen because one of the members of the object
+            # may not be serializable for cloudpickle. So we need
+            # these extra fallbacks here to start from the beginning.
+            # Hopefully the object could have a `__reduce__` method.
+            register_custom_serializer(type(value), use_pickle=True)
+            warning_message = ("WARNING: Serializing the class {} failed, "
+                               "falling back to cloudpickle.".format(
+                                   type(value)))
+            logger.warning(warning_message)
+            self.store_and_register(object_id, value)
 
     def retrieve_and_deserialize(self, object_ids, timeout, error_timeout=10):
         start_time = time.time()
